let's gooooo

+++ test/Procs.dm
@@ -1,7 +1,17 @@
 
 proc/editableLists()
 	return world.contents
+var/worldtime = 1
 
+proc/Save_Gain()
+	var/savefile/S=new("GAIN")
+	S["GAIN"]<< globalgain
+	S["Time"]<<worldtime
+proc/Load_Gain()
+	if(fexists("GAIN"))
+		var/savefile/S=new("GAIN")
+		S["GAIN"]>> globalgain
+		S["Time"]>>worldtime
 mob
 	proc
 		TakeDamage(var/Damage,var/mob/Attacker)
@@ -25,37 +35,63 @@ mob
 			view() << "[fyellow] [src] [fend] is KOED!"
 			icon_state = "KO"
 			src.can_move = 0
-			sleep(120)
+			sleep(1200)
 			view() << "[fgreen] [src] [fend] stands up again!"
 			src.ko = 0
 			src.can_move = 1
 			src.canattack = 1
 		KillCheck(var/mob/M)
 			if (M.ko == 1)
-				switch(alert("Do you want to kill them?",,"Yes","No"))
-					if("Yes")
-						if(M.client)
-							world<<"[src] killed [M]!"
-							M.life=M.maxlife
-							flick("Fall",M)
-							spawncorpse(M)
-							M.loc=locate(/turf/start)
-						else
-							src<<"<b>You killed [M]."
-							src.plgain()
-							flick("Fall",M)
-							spawncorpse(M)
-							del M
+				if(src.client)
+					switch(alert("Do you want to kill them?",,"Yes","No"))
+						if("Yes")
+							if(M.client)
+								world<<"[src] killed [M]!"
+								M.life=M.maxlife
+								flick("Fall",M)
+								spawncorpse(M)
+								M.dead = 1
+								Death(M)
+								M.ko = 0
+								M.can_move = 1
+								M.canattack = 1
+								M.loc=locate(/turf/start)
+							else
+								src<<"<b>You killed [M]."
+								src.plgain()
+								flick("Fall",M)
+								spawncorpse(M)
+								M.dead = 1
+								del M
 
-					if("No")
-						view() << "[src] gave mercy on [M]"
-						M.loc = M.loc
+						if("No")
+							view() << "[src] gave mercy on [M]"
+							M.loc = M.loc
+				else
+					Death(M)
+					flick("Fall",M)
+					M.life=M.maxlife
+					spawncorpse(M)
+					M.dead=1
+					M.ko = 0
+					M.can_move = 1
+					M.canattack = 1
 		DeathCheck(var/mob/Killer)
 			if(src.life<=0)
 				KO(src)
 			if(src.ko == 1)
 				view() << "[Killer] is about to kill [src]"
-
+mob/var/firstdeath=1
+mob
+	proc
+		Death()
+			if (player&&dead)
+				if(firstdeath)
+					firstdeath=0
+					var/DeathPL = gains*1000*plmod
+					if(DeathPL>recordpl) DeathPL = recordpl
+					maxpl += DeathPL
+					src.loc = /turf/start
 		knockback(var/r,var/d) // r range, d dir
 			if(!r) return
 			while(r > 0)
@@ -69,7 +105,7 @@ mob
 				dir=turn(dir,180)
 			if(Bump(null))
 				return
-			if(Bump(/obj/wall))
+			if(Bump(/obj/destructable))
 				new/obj/dust(src.loc,src.dir)
 				new/obj/dust(src.loc,src.dir)
 		stamcheck()
@@ -85,16 +121,16 @@ mob/proc/spawncorpse()
 	S.desc = "[src]'s corpse. May \he rest in peace."
 	S.name = "[src]'s corpse"
 
-obj/wall
+obj/destructable
 	proc
 		Decheck(var/Damage)
 			durability -= Damage
 			if(src==null) return 1
-			if(durability < 100 && durability >= 50)
+			if(durability < maxdurability && durability >= (maxdurability/2))
 				src.overlays += "overlay_damage"
-			if(durability < 50 && durability >= 25)
+			if(durability < (maxdurability/2) && durability >= (maxdurability/4))
 				src.overlays += "overlay_damage2"
-			if(durability < 25 && durability >= 1)
+			if(durability < (maxdurability/4) && durability >= 1)
 				src.overlays += "overlay_damage3"
 			if(durability <= 0)
 				density=0
@@ -111,10 +147,11 @@ mob
 				sleep(15)
 				plgain(0.5)
 				energygain(0.1)
+				speedgain(0.5)
 				statgain("fight",0.3)
 				src.energy -= src.maxenergy / (10*src.maxenergy)
 				src.lift += rand(3,5)*((src.strengthmod + src.defensemod)/200)
-				ratinggain()
+				ratinggain(1)
 				if(src.energy <= 0)
 					spawn(5)stamcheck()
 					src << "You stop training due to exhaustion."
@@ -128,8 +165,11 @@ mob
 			src.can_move = 1
 			src.canattack = 1
 			src << "You stop training."
-		ratinggain()
-			src.rating += round(rand(1,5)*(src.rating/100))
+		ratinggain(mod)
+			if (src.rating <= 100)
+				src.rating+= rand(1,50)
+			else
+				src.rating += mod*round(rand(1,5)*(src.rating/10000))
 			sleep(15)
 	proc
 		MeditatingLoop()
@@ -139,7 +179,11 @@ mob
 				plgain(0.3)
 				src.energy ++
 				energygain(0.5)
+				ratinggain(0.5)
+				speedgain(1)
 				statgain("spirit",0.3)
+				if (inbed && (tired <= 100))
+					tired+=1
 				spawn(5)src.stamcheck()
 				if (src.energy >= src.maxenergy) src.energy = src.maxenergy
 				if(src.meditating==0)
@@ -151,15 +195,18 @@ mob
 			src.can_move = 1
 			src.canattack = 1
 
+		speedgain(gain)
+			src.speed += globalgain*gain*speedmod*(src.rating/100)*0.001
 		plgain(gain)
-			src.maxpl += src.globalgain*gain*1*10 * (src.rating/100)
+			src.maxpl += globalgain*gain*1*plmod*10 * (src.rating/100)
 		energygain(gain)
-			src.maxenergy += src.globalgain*gain*forcemod*10 *(src.rating/100)
+			src.maxenergy += globalgain*gain*forcemod*10 *(src.rating/100)
 		statgain(focus,gain)
 			switch(focus)
 				if ("spirit")
-					src.force+= src.globalgain*gain*forcemod*50 *(src.rating/100)
-					src.resistance+= src.globalgain*gain*resmod*50 *(src.rating/100)
+					src.force+= globalgain*gain*forcemod *(src.rating/100)
+					src.resistance+= globalgain*gain*resmod *(src.rating/100)
 				if("fight")
-					src.defense+= src.globalgain*gain*defensemod*50 *(src.rating/100)
-					src.strength+=src.globalgain*gain* strengthmod*50 *(src.rating/100)
+					src.defense+= globalgain*gain*defensemod *(src.rating/100)
+					src.strength+=globalgain*gain* strengthmod *(src.rating/100)
+					src.lift += rand(3,5)*((src.strengthmod + src.defensemod)/200*(src.rating/100))
